+++ common/src/obj.rs
@@ -44,13 +44,13 @@ pub struct CharaTemplateObject {
     /// Default AI kind for this character
     pub default_ai_kind: gamedata::NpcAIKind,
     pub max_hp: i32,
-    pub str: u32,
-    pub vit: u32,
-    pub dex: u32,
-    pub int: u32,
-    pub wil: u32,
-    pub cha: u32,
-    pub spd: u32,
+    pub str: u16,
+    pub vit: u16,
+    pub dex: u16,
+    pub int: u16,
+    pub wil: u16,
+    pub cha: u16,
+    pub spd: u16,
 }
 
 #[derive(Serialize, Deserialize)]
